+++ core/silkworm/common/util.hpp
@@ -22,8 +22,8 @@
 #endif
 
 #include <cstring>
-#include <optional>
 #include <ethash/keccak.hpp>
+#include <optional>
 #include <silkworm/common/base.hpp>
 
 namespace silkworm {
@@ -78,7 +78,7 @@ Bytes from_hex(std::string_view hex);
 
 // Parses a string input value representing a size in
 // human readable format with qualifiers. eg "256MB"
-std::optional<size_t> parse_size(const std::string& sizestr);
+std::optional<uint64_t> parse_size(const std::string& sizestr);
 
 // TODO[C++20] replace by starts_with
 inline bool has_prefix(ByteView s, ByteView prefix) { return s.substr(0, prefix.size()) == prefix; }
